# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Talent Match API",
        "description": "Back End API Using Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "LOGIN",
                "description": "LOGIN THE USERS FROM DATABASE",
                "operationId": "LOGIN_ID",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "REGISTER USER",
                "description": "REGISTER USER TO THE USERS FROM DATABASE",
                "operationId": "REGISTER_ID",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "GET ALL USER",
                "description": "DELETE USERS FROM DATABASE",
                "operationId": "LOGOUT_ID",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
...
        "/api/job_post": {
            "get": {
                "tags": [
                    "Job_post"
                ],
                "summary": "list of all job_post",
                "description": "this API list all information about job_posts",
                "operationId": "index_job_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/job_post/store": {
            "post": {
                "tags": [
                    "Job_post"
                ],
                "summary": "store_job_post info",
                "description": "this API store information of job_posts",
                "operationId": "store_job_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
...
    }
}